Fix broken links, a couple of typos
A number of the internal links weren't formatted correctly, so didn't
work.

Likewise a couple of <a name=""> declarations were incorrectly <a
href="">.

Some external http links were auto-redirecting to https, so fix those
too.

Escape one pair of <>

A couple of typos that don't change the meaning of the text also
fixed.

Signed-off-by: Matthew Vernon <mv3@sanger.ac.uk>